requirements.txt
flask
python-dotenv
py-money
gunicorn
flask-talisman
